requirements.txt
streamlit>=1.40.0
fastapi>=0.115.0
uvicorn>=0.30.0
chromadb>=0.4.22
httpx>=0.27.0
langchain-google-genai>=4.0.0
apscheduler>=3.10.4
beautifulsoup4>=4.12.0
lxml>=5.1.0
sentence-transformers>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.6.0